markdowndocs

+++ Jumpscale/clients/sshagent/SSHAgent.py
@@ -8,6 +8,9 @@ class SSHAgent(j.application.JSFactoryBaseClass):
 
     __jslocation__ = "j.clients.sshagent"
 
+    def _init(self):
+        self._available=None
+
     @property
     def keyname_default(self):
         """
@@ -236,24 +239,30 @@ class SSHAgent(j.application.JSFactoryBaseClass):
         Check if agent available
         :return: bool
         """
-        socket_path = self.ssh_socket_path
-        if not j.sal.fs.exists(socket_path):
-            return False
-        if "SSH_AUTH_SOCK" not in os.environ:
-            self._init_ssh_env()
-        return_code, out, _ = j.sal.process.execute("ssh-add -l",
-                                                    showout=False,
-                                                    die=False, useShell=False)
-        if 'The agent has no identities.' in out:
-            return True
+        if self._available is None:
+            socket_path = self.ssh_socket_path
+            if not j.sal.fs.exists(socket_path):
+                self._available = False
+                return False
+            if "SSH_AUTH_SOCK" not in os.environ:
+                self._init_ssh_env()
+            return_code, out, _ = j.sal.process.execute("ssh-add -l",
+                                                        showout=False,
+                                                        die=False, useShell=False)
+            if 'The agent has no identities.' in out:
+                self._available = True
+                return True
 
-        if return_code != 0:
-            # Remove old socket if can't connect
-            if j.sal.fs.exists(socket_path):
-                j.sal.fs.remove(socket_path)
-            return False
-        else:
-            return True
+            if return_code != 0:
+                # Remove old socket if can't connect
+                if j.sal.fs.exists(socket_path):
+                    j.sal.fs.remove(socket_path)
+                self._available = False
+                return False
+            else:
+                self._available = True
+                return True
+        return self._available
 
     def kill(self, socketpath=None):
         """
